Allocate all the distillation qubits for Clifford correction
Before this change, only one qubit in the distillation area was
selected for potential Clifford correction. This was incorrect
because using the qubit for Clifford correction was not possible,
depending on the distillation results.

This change resolves the issue by allocating all qubits in the
distillation area for potential use in Clifford correction.